Using getMy.lua to hold branch preference

+++ turtleApps/src/getMy.lua
@@ -11,7 +11,9 @@ at http://opensource.org/licenses/MIT)
 ]]
 
 -- NOTE on turtle copy, update branch
-local MY_BRANCH= "master/"
+MY_BRANCH= "master/"
+
+DEP_VERSION="1.1"
 
 local MY_BASE = "https://"..
   "raw.githubusercontent.com/"..
@@ -21,22 +23,26 @@ local MY_BASE = "https://"..
   "turtleApps/src/"
 
 local args= {...}
-local lclFile= args[1]
-local lclOldFile= "OLD_".. lclFile
 
-if fs.exists( lclOldFile ) then
-  fs.delete( lclOldFile )
-  print("deleted ".. lclOldFile)
+if table.getn( args ) > 0 then
+  local lclFile= args[1]
+  local lclOldFile= "OLD_".. lclFile
+  
+  if fs.exists( lclOldFile ) then
+    fs.delete( lclOldFile )
+    print("deleted ".. lclOldFile)
+  end
+  
+  if fs.exists( lclFile ) then
+    fs.move( lclFile, lclOldFile )
+    print("Renamed original to ".. 
+        lclOldFile )
+  end
+  
+  local url= MY_BASE.. lclFile.. ".lua"
+  
+  shell.run("wget", url, lclFile)
+  
+  print("Got a new copy of ".. lclFile)
+
 end
-
-if fs.exists( lclFile ) then
-  fs.move( lclFile, lclOldFile )
-  print("Renamed original to ".. 
-      lclOldFile )
-end
-
-local url= MY_BASE.. lclFile.. ".lua"
-
-shell.run("wget", url, lclFile)
-
-print( "Got a new copy of ".. lclFile)
